fix(gpu): prevent reshape axis integer overflow

+++ src/plugins/intel_gpu/src/graph/graph_optimizer/prepare_buffer_fusing.cpp
@@ -740,7 +740,7 @@ void crop_in_place_optimization::update_in_place_crop_padding_along_feature(cons
                 // axis-mapping so build-time (dyn mask) and runtime (mask +
                 // explicit sizes) paths pick the same reshape axis for patterns
                 // that do not need scaling.
-                size_t reshape_axis = crop_axis;
+                int64_t reshape_axis = static_cast<int64_t>(crop_axis);
                 if (reshape_mode == reshape::reshape_mode::base) {
                     reshape_axis = reshape_ps.size() - 1;
                     ov::Dimension::value_type mul = 1;
@@ -753,7 +753,7 @@ void crop_in_place_optimization::update_in_place_crop_padding_along_feature(cons
                 } else if (reshape_mode == reshape::reshape_mode::unsqueeze || reshape_mode == reshape::reshape_mode::squeeze) {
                     const auto& output_pattern = reshape_desc->output_pattern;
                     for (size_t i = 0; i < output_pattern.size(); i++) {
-                        if (output_pattern[i] <= static_cast<int64_t>(reshape_axis)) {
+                        if (output_pattern[i] <= reshape_axis) {
                             reshape_axis += reshape_mode == reshape::reshape_mode::unsqueeze ? 1 : -1;
                         }
                     }
